+++ generatedTypes/incubator/index.d.ts
@@ -2,3 +2,4 @@ export { default as TabController } from './TabController';
 export { default as TextField, TextFieldProps, FieldContextType } from './TextField';
 export { default as TouchableOpacity, TouchableOpacityProps } from './TouchableOpacity';
 export { default as WheelPicker, WheelPickerProps } from './WheelPicker';
+export { default as PanView, PanViewProps, PanningDirections, PanDismissThreshold } from './panView';
